Reduce PR diff
